% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{add_all_hypotheticals}
\alias{add_all_hypotheticals}
\title{Add All Hypothetical Parent Entries to the Studbook}
\usage{
add_all_hypotheticals(studbook, hyp_defs, loc_key)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{hyp_defs}{A list containing definitions for individuals that remain missing a parental assignment after manual matching using \code{find_parent}.}

\item{loc_key}{A data frame containing location key mappings produced by \code{read_locations}.}
}
\value{
A data frame with all hypothetical parent entries added as new rows and assigned to the relevant offspring.
}
\description{
Combines hypothetical parent entries for both sires and dams and appends them to the studbook. This function processes a list of missing parentage definitions and returns an updated studbook with new hypothetical entries.
}
