% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{age_classes_ids}
\alias{age_classes_ids}
\title{Generate a tibble with one row per year of life for each studbook ID born within a demographic window}
\usage{
age_classes_ids(studbook, window = NULL)
}
\arguments{
\item{window}{vector of start and end date (must be provided as \code{ymd()} values) to use as the demographic window for calculations
(default is the date of the first captive birth recorded and the most recent date of birth recorded)}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with one row per year of life for each studbook ID so that the \code{nrow()} per id matches their age in years at death (or current age, if living)
Note that \code{x} represents the age in years, which is calculated based on start of annual quarters so that an individuals date of birth is rolled back to the start of a given annual quarter,
and \code{x_start} provides the date of that individual's birth quarter and \code{x_end} provides the last date within that age interval for the individual.
}
\description{
Generate a tibble with one row per year of life for each studbook ID born within a demographic window
}
