% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{animate_generations}
\alias{animate_generations}
\title{Add animation to age-specific plots over multiple generations}
\usage{
animate_generations(plot, data, attr)
}
\arguments{
\item{plot}{plotly object to animate}

\item{data}{A dataframe to use in plotly (generated internally using the function \code{plot_demog_prep})}

\item{attr}{A list of attribute values created using the internal function \code{plot_demog_attr}}
}
\value{
A list of parameter values to use in plotly plots of demographic variables
}
\description{
Add animation to age-specific plots over multiple generations
}
