% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{births_annual}
\alias{births_annual}
\title{Count births per individual per year}
\usage{
births_annual(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with birth counts per studbook ID per year
}
\description{
Count births per individual per year
}
