% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{build_ped_network}
\alias{build_ped_network}
\title{Generate nodes and edges from pedigree for a network object}
\usage{
build_ped_network(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of edges with visual attributes
}
\description{
Generate nodes and edges from pedigree for a network object
}
