% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{build_ped_series}
\alias{build_ped_series}
\title{Create a List of Pedigree Objects from the Studbook}
\usage{
build_ped_series(studbook)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{\link[=read_studbook]{read_studbook()}}.}
}
\value{
A list of pedigree objects excluding any singletons.
}
\description{
This function converts a studbook into a list of pedigree objects (using
\code{pedtools::ped()}) and excludes any singletons (pedigrees with a size of 1 or less).
}
