% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate Age}
\usage{
calculate_age(birth, date)
}
\arguments{
\item{birth}{A date object representing the birth date.}

\item{date}{A date object representing the reference date.}
}
\value{
An integer representing the age in whole years.
}
\description{
Computes the age in whole years based on a given birth date and a reference date.
}
