% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{caption_plotly}
\alias{caption_plotly}
\title{Add a formatted figure caption to plotly plots}
\usage{
caption_plotly(caption, number)
}
\arguments{
\item{caption}{Caption text to add}

\item{number}{Figure number to apply to caption (optional)}
}
\value{
nested list to add to plotly layout function
}
\description{
Add a formatted figure caption to plotly plots
}
