% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{census}
\alias{census}
\title{Generate census summary by sex over time}
\usage{
census(studbook, period)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{period}{"years" or "months"}
}
\value{
A wide-format census summary with counts per sex and date
}
\description{
Generate census summary by sex over time
}
