% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{census_pop}
\alias{census_pop}
\title{Count individuals by age and cohort}
\usage{
census_pop(studbook, span = 5, annual = FALSE, by_age = TRUE)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{annual}{logical indicating whether to sum counts from each age class by year (or to sum across multi-year birth cohorts)}

\item{by_age}{logical indicating whether to count by age class or across all ages}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with population counts
}
\description{
Count individuals by age and cohort
}
