% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{choose_pedigree}
\alias{choose_pedigree}
\title{Choose a Pedigree Based on Maximum Depth}
\usage{
choose_pedigree(pedigree_series)
}
\arguments{
\item{pedigree_series}{A list of pedigree objects.}
}
\value{
A character string representing the selected pedigree name.
}
\description{
This function selects the pedigree name corresponding to the highest computed depth
from a series of pedigrees.
}
