% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohort_births}
\alias{cohort_births}
\title{Count births for each individual across each of their age classes}
\usage{
cohort_births(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with x rows per ID (with x being each age class integer) birth counts in columns
}
\description{
Count births for each individual across each of their age classes
}
