% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{cohort_defaults}
\alias{cohort_defaults}
\title{Generate a list of default parameters for setting up cohort tibbles}
\usage{
cohort_defaults(studbook, cohort_params = list(NULL, ...))
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
A list of named values to use for cohort defaults in other functions
}
\description{
Generate a list of default parameters for setting up cohort tibbles
}
