% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{cohort_demog}
\alias{cohort_demog}
\title{Generate demography summary of studbook data over cohorts}
\usage{
cohort_demog(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
A tibble with cohort life table summary
}
\description{
Generate demography summary of studbook data over cohorts
}
