% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohort_lifelist}
\alias{cohort_lifelist}
\title{Nest life table data into vectors listed by birth cohort}
\usage{
cohort_lifelist(lifetab)
}
\arguments{
\item{lifetab}{tibble generated using the function \code{cohort_lifetab}}
}
\value{
nested list with each birth cohort at the top level (name represents the cohort's minimum birth year),
and each of those lists containing \code{x} (age class), \code{Sx} (population count),
\code{nLx} (person-years lived in the interval \code{x} to \code{x+n}),
\code{lx} (cumulative survivorship), \code{mx} (per-capita birth rate)
}
\description{
Nest life table data into vectors listed by birth cohort
}
