% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohort_lifetab}
\alias{cohort_lifetab}
\title{Use \code{demogR} to create a list of life tables for each cohort}
\usage{
cohort_lifetab(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with one row per birth cohort and tibble-columns \code{lt} (tibble generated by \code{demogR::life.table}),
\code{life} (life table results joined to basic vital stats from \code{data} column produced by \code{vitals}),
as well as a nested list of vectors,  \code{les}, for computing leslie matrices.
}
\description{
Use \code{demogR} to create a list of life tables for each cohort
}
