% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohort_studbook}
\alias{cohort_studbook}
\title{Match birth cohorts to studbook IDs}
\usage{
cohort_studbook(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with each studbook ID matched to a birth cohort
}
\description{
Match birth cohorts to studbook IDs
}
