% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohort_template}
\alias{cohort_template}
\title{Set up cohorts by birth year for population status}
\usage{
cohort_template(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble with \code{cohort} (numeric id), \code{cohort_start} (earliest birth year of cohort), \code{cohort_end} (last birth year of cohort),
\code{born} (hypothetical birth year), \code{sex} (hypothetical sex), \code{x} (age in years), \code{x_n} (x + n for lifetable)
}
\description{
Set up cohorts by birth year for population status
}
