% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohorts_leslie}
\alias{cohorts_leslie}
\title{Create a leslie matrix for each birth cohort using \code{demogR::leslie.matrix}}
\usage{
cohorts_leslie(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
nested list with each birth cohort at the top level (name represents the cohort's minimum birth year),
and each of those lists containing a leslie matrix, or an error message if data were insufficient for computation
}
\description{
Create a leslie matrix for each birth cohort using \code{demogR::leslie.matrix}
}
