% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{cohorts_set}
\alias{cohorts_set}
\title{Set up cohorts by birth year for population status}
\usage{
cohorts_set(studbook, span = 5)
}
\arguments{
\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
List of tibbles to organize data into birth cohorts
}
\description{
Set up cohorts by birth year for population status
}
