% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/lifetab.R
\name{count_births}
\alias{count_births}
\title{Count births and track individual presence over time}
\usage{
count_births(studbook, span = 5)

count_births(studbook, span = 5)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{span}{integer representing the width of each birth-cohort in years}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
A tibble of births per ID by year

Tibble with x rows per ID (with x being each age class integer) birth counts in columns
}
\description{
Count births and track individual presence over time

Count births for each individual across each of their age classes
}
