% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog.R
\name{demog_summary}
\alias{demog_summary}
\title{Reduce full life table to summary format for visualization}
\usage{
demog_summary(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
Nested summary table with stats calculated by age in list-cols
}
\description{
Reduce full life table to summary format for visualization
}
