% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog.R
\name{demog_summary_total}
\alias{demog_summary_total}
\title{Reduce full life table to summary format for the population across all years and cohorts}
\usage{
demog_summary_total(studbook)
}
\arguments{
\item{studbook}{Studbook tibble}
}
\value{
Nested, condensed summary table with stats calculated by age in list-cols
}
\description{
Reduce full life table to summary format for the population across all years and cohorts
}
