% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog.R
\name{demog_tab}
\alias{demog_tab}
\title{Build a full demographic life table from cohort data}
\usage{
demog_tab(df)
}
\arguments{
\item{df}{A cohort-formatted tibble with Births, Nx, Age, Cohort}
}
\value{
A life table with survivorship, mortality, and reproductive values
}
\description{
Build a full demographic life table from cohort data
}
