% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{demog_variables}
\alias{demog_variables}
\title{Generate a list with descriptions for values calculated by demographic functions}
\usage{
demog_variables()
}
\value{
nested list to annotate visuals
}
\description{
Generate a list with descriptions for values calculated by demographic functions
}
