% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{demog_wide}
\alias{demog_wide}
\title{Pivot a life table wider for ease of plotting multiple traces}
\usage{
demog_wide(df, group, x, y)
}
\arguments{
\item{df}{A data frame with demography stats for plotting}

\item{group}{The grouping variable(s) to plot as individual traces}

\item{x}{The variable(s) that will be plotted on the x axis}

\item{y}{The variable(s) that will be plotted on the y axis}
}
\value{
Wider version with sexes in columns for plotting multiple traces
}
\description{
Pivot a life table wider for ease of plotting multiple traces
}
