% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{demog_ylab}
\alias{demog_ylab}
\title{Generate formatted y-axis labels for demographic variables}
\usage{
demog_ylab(variable)
}
\arguments{
\item{variable}{Name of the column containing an age-specific variable to map onto the y-axis
options for variable: \code{N0}, \code{N1}, \code{R0}, \code{T}, \code{MLE}, \code{Repro_first}, \code{Repro_last}, \code{age_max}, \code{lambda}, \code{r},
\code{Births}, \code{Deaths}, \code{Nx}, \code{Qx}, \code{Lx}, \code{Lx1}, \code{Px}, \code{ex}, \code{Tx}, \code{Mx}, \code{Fx}}
}
\value{
nested list of strings to use for labeling y-variables
}
\description{
Generate formatted y-axis labels for demographic variables
}
