% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{display_pos_parents}
\alias{display_pos_parents}
\title{Generate a shiny-ui element that displays possible parentage assignments when constructing a new studbook}
\usage{
display_pos_parents(list, parent = NULL)
}
\arguments{
\item{list}{A list of tibbles containing matched parent details. Each tibble corresponds to an individual and lists potential parent matches.
Create this list using the \code{find_parent} function}

\item{parent}{A string to label the output with which parent is represented (optional)}
}
\value{
Shiny html object that renders as collapsed accordion panels
}
\description{
Generate a shiny-ui element that displays possible parentage assignments when constructing a new studbook
}
