% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_children}
\alias{edges_children}
\title{Create Edges Connecting Parents to Children}
\usage{
edges_children(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges linking parents with their children.
}
\description{
This function builds an edge data frame linking parents to children in the pedigree,
with associated connector colors.
}
