% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_dads}
\alias{edges_dads}
\title{Create Edges for Paternal Connections}
\usage{
edges_dads(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges representing paternal connections.
}
\description{
This function extracts the edge data that connects fathers in the pedigree.
}
