% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_hubs}
\alias{edges_hubs}
\title{Create Edges Connecting Parents to Children}
\usage{
edges_hubs(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges for subnuclei connections.
}
\description{
This function creates an edge data frame linking parent connector nodes to offspring connector nodes in the pedigree
}
