% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{edges_igraph}
\alias{edges_igraph}
\title{Create Edge Data for an igraph Pedigree Graph}
\usage{
edges_igraph(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges with attributes for igraph rendering.
}
\description{
This function constructs an edge data frame for building an igraph object using pedigree data.
}
