% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_kids}
\alias{edges_kids}
\title{Create Edges Connecting terminal child nodes in pedigree}
\usage{
edges_kids(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges for mom connections.
}
\description{
This function creates an edge data frame linking offspring connector nodes to each child node
}
