% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_links}
\alias{edges_links}
\title{Create Edges connecting an individual across subnuclei/transfers}
\usage{
edges_links(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges for subnuclei connections.
}
\description{
This function creates an edge data frame linking any individual across multiple node representations
(Some individuals are a part of multiple subnuclear units based on transfers over time)
}
