% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_moms}
\alias{edges_moms}
\title{Create Edges for Maternal Connections}
\usage{
edges_moms(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges representing maternal connections.
}
\description{
This function extracts the edge data that connects mothers in the pedigree.
}
