% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{edges_subnucs}
\alias{edges_subnucs}
\title{Create Edges for Subnuclear Connections}
\usage{
edges_subnucs(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges for subnuclei connections.
}
\description{
This function creates an edge data frame linking subnuclei nodes in the pedigree,
and then merges these with child connection data.
}
