% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{est_date_btn}
\alias{est_date_btn}
\title{Estimate the Midpoint Date Between Two Dates}
\usage{
est_date_btn(date1, date2)
}
\arguments{
\item{date1}{A date object representing the start date.}

\item{date2}{A date object representing the end date.}
}
\value{
A date object corresponding to the midpoint between \code{date1} and \code{date2}.
}
\description{
Computes an estimated date that is approximately halfway between \code{date1} and \code{date2}.
}
