% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{extract_pedigree}
\alias{extract_pedigree}
\title{Extract a pedigree of interest as its own object}
\usage{
extract_pedigree(pedigree_series, name, title = NULL)
}
\arguments{
\item{pedigree_series}{A list of pedigree objects.}

\item{name}{The name of the pedigree object to extract}

\item{title}{The new name to apply to the extracted pedigree (optional)}
}
\value{
A single pedigree object representing the living population
}
\description{
For many of our operations, it will be easier to work with one pedigree object of interest and ignore the pedigrees not currently represented in the breeding population. The function below helps you extract the pedigree of interest as its own object with a new name.
}
