% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fam_tooltip}
\alias{fam_tooltip}
\title{Create a character vector from a studbook dataframe to represent the connector between parents and offspring in a pedigree}
\usage{
fam_tooltip(df, name)
}
\arguments{
\item{df}{A data frame with rowwise location data containing the following columns:
\code{dad}, \code{mom}, \code{iconLoc}, and \code{Loc}}

\item{name}{A string to use as the name of the new column}
}
\value{
The same data frame with a new string column
}
\description{
Create a character vector from a studbook dataframe to represent the connector between parents and offspring in a pedigree
}
