% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{featured_pedigree}
\alias{featured_pedigree}
\title{Get the Featured Pedigree from a Studook}
\usage{
featured_pedigree(studbook)
}
\arguments{
\item{studbook}{A data frame containing studbook data.}
}
\value{
A pedigree object representing the featured pedigree.
}
\description{
This function builds the pedigree series from the studbook and extracts the featured pedigree.
}
