% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{find_parent}
\alias{find_parent}
\title{Find Parent Information from the Studbook}
\usage{
find_parent(studbook, loc_key, parent)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{loc_key}{A data frame containing location key mappings produced by \code{read_locations}.}

\item{parent}{A character string indicating the parent type (e.g., "Sire", "sire", "dad", "father" for males; analogous values for females).}
}
\value{
A list of tibbles containing matched parent details. Each tibble corresponds to an individual and lists potential parent matches.
}
\description{
Searches the studbook for potential parent candidates (sire or dam) based on the birth institution and birth year, and returns matching details. This function assists in identifying likely paternity or maternity assignments.
}
