% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_demog.R
\name{foot_demog}
\alias{foot_demog}
\title{Create a list of values to use as a footer row in a reactable summary of demographic variables}
\usage{
foot_demog(studbook)
}
\arguments{
\item{studbook}{Studbook tibble}
}
\description{
Create a list of values to use as a footer row in a reactable summary of demographic variables
}
