% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{founder.cols}
\alias{founder.cols}
\title{Column definitions for founder summary tables}
\usage{
founder.cols(df)
}
\arguments{
\item{df}{A filtered founder summary dataframe}
}
\value{
A named list of column definitions
}
\description{
Column definitions for founder summary tables
}
