% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_numbers}
\alias{gen_numbers}
\title{Return number of generations for individuals}
\usage{
gen_numbers(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object from \code{pedtools::ped()}}
}
\value{
Named vector of generation numbers
}
\description{
Return number of generations for individuals
}
