% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_numbers_living}
\alias{gen_numbers_living}
\title{Get generation numbers for living individuals}
\usage{
gen_numbers_living(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object from \code{pedtools::ped()}}

\item{studbook}{Studbook tibble}
}
\value{
Named vector of generations for living IDs
}
\description{
Get generation numbers for living individuals
}
