% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hline}
\alias{hline}
\title{Add a horizontal reference line to plotly layout}
\usage{
hline(y = 0, color = "#444444")
}
\arguments{
\item{y}{Numeric y-intercept for the line}

\item{color}{Line color as a hex string}
}
\value{
A plotly-compatible line specification
}
\description{
Add a horizontal reference line to plotly layout
}
\keyword{internal}
