% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{igraph_edges}
\alias{igraph_edges}
\title{Create a pedigree edge dataframe that will work with \code{igraph}}
\usage{
igraph_edges(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
A dataframe of edges to use with \code{igraph}
}
\description{
Create a pedigree edge dataframe that will work with \code{igraph}
}
