% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{kin.cols}
\alias{kin.cols}
\title{Column definitions for kinship summary tables}
\usage{
kin.cols()
}
\value{
A named list of column definitions
}
\description{
Column definitions for kinship summary tables
}
