% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{label_names}
\alias{label_names}
\title{Label Names}
\usage{
label_names(studbook, names = NULL)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{names}{Factor labeling vector written as \verb{c(\{String\} = "\{ID\}")} for assigning each name to an ID number.}
}
\value{
A data frame of studbook data now containing a column \code{label_special} with a string only for those IDs specified in the \code{names} parameter.
}
\description{
Adds names or other string values to a column based on unique, numerical IDs
}
