% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{lifeTab_static}
\alias{lifeTab_static}
\title{Reduce full life table to summary stats per cohort}
\usage{
lifeTab_static(df)
}
\arguments{
\item{df}{Life table tibble}
}
\value{
Condensed summary table with lambda and vital rates
}
\description{
Reduce full life table to summary stats per cohort
}
