% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{lifetime_tab}
\alias{lifetime_tab}
\title{Reduce full life table to summary stats per cohort}
\usage{
lifetime_tab(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
Condensed summary table with lambda and vital rates
}
\description{
Reduce full life table to summary stats per cohort
}
