% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog.R
\name{lifetime_total}
\alias{lifetime_total}
\title{Reduce full life table to summary stats for the population across all years and cohorts}
\usage{
lifetime_total(studbook)
}
\arguments{
\item{studbook}{Studbook tibble}
}
\value{
Condensed summary table with lambda and vital rates (3 rows - Females, Males, and Overall stats)
}
\description{
Reduce full life table to summary stats for the population across all years and cohorts
}
