% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{lighten_palette}
\alias{lighten_palette}
\title{Lighten color palette by replacing hex alpha value}
\usage{
lighten_palette(palette, hex)
}
\arguments{
\item{palette}{A named list or vector of colors}

\item{hex}{A two-character hex string (e.g., "33")}
}
\value{
Palette with adjusted transparency
}
\description{
Lighten color palette by replacing hex alpha value
}
