% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{living.cols}
\alias{living.cols}
\title{Column definitions for living individual summaries}
\usage{
living.cols(df)
}
\arguments{
\item{df}{A living population summary dataframe}
}
\value{
A named list of column definitions
}
\description{
Column definitions for living individual summaries
}
