% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loc_tooltip}
\alias{loc_tooltip}
\title{Create a character vector from a studbook dataframe to represent location data in a tooltip}
\usage{
loc_tooltip(df)
}
\arguments{
\item{df}{A data frame with rowwise location data containing the columns \code{iconLoc_birth}, \code{Loc_birth}, \code{Institution_birth}, and \code{State_Province_birth}}
}
\value{
The same data frame with a new string column
}
\description{
Create a character vector from a studbook dataframe to represent location data in a tooltip
}
