% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{location_census}
\alias{location_census}
\title{Generate Location Census Data}
\usage{
location_census(studbook, loc_key)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{loc_key}{A data frame containing location key mappings produced by \code{read_locations}.}
}
\value{
A nested list of location census data organized by year.
}
\description{
Creates an annual census of locations based on studbook data and location key mappings. The result is a nested list organized by the floor date (January 1) for each year represented in the data. Each year contains a list of locations, with each location represented by a tibble listing the individuals recorded there along with their sex and age.
}
