% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_cohorts}
\alias{make_cohorts}
\title{Generate cohort-structured data}
\usage{
make_cohorts(studbook, Year_min, Year_max, span, age_max, include_sex = TRUE)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{Year_min}{Start year of cohorts}

\item{Year_max}{End year}

\item{span}{Years per cohort}

\item{age_max}{Max age to include}

\item{include_sex}{Whether to include sex as a grouping var}
}
\value{
A joined and restructured tibble
}
\description{
Generate cohort-structured data
}
