% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{make_leslie}
\alias{make_leslie}
\title{Create a Leslie Matrix using demographic summary table}
\usage{
make_leslie(F, S)
}
\arguments{
\item{F}{name of column containing fecundities}

\item{S}{name of column containing survival probabilities}
}
\value{
Values for matrix \code{A}
}
\description{
Create a Leslie Matrix using demographic summary table
}
