% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{mortality_history}
\alias{mortality_history}
\title{Generate a summary of deaths across population history}
\usage{
mortality_history(studbook, window = NULL)
}
\arguments{
\item{window}{vector of start and end date (must be provided as \code{ymd()} values) to use as the demographic window for calculations
(default is the date of the first captive birth recorded and the most recent date of birth recorded)}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble summarizing the deaths across studbook records within the demographic window
(\code{ID} of deceased individual, \code{Sex} of individual, \code{Type_birth} indicating \code{Wild} for wild captures or
\code{Captive} for captive-born individuals, \code{Qtr_born} is the date used to represent birth dates - which are converted to the starting date of the annual quarter,
\code{Qtr_death}is the date used to represent death dates - which are converted to the starting date of the annual quarter, \code{x_death} is the approximate age class in years
in which the individual dies based on the difference between the start of the quarter of their birth and the start of the quarter of their death)
}
\description{
Generate a summary of deaths across population history
}
