% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{nodes_connectors}
\alias{nodes_connectors}
\title{Create Nodes for Subnuclear Connections}
\usage{
nodes_connectors(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of nodes for subnuclei connections.
}
\description{
This function creates a template node data frame linking subnuclei nodes in the pedigree.
(variables can be selected or removed later depending on which visualization tool you use)
}
