% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{nodes_subjects}
\alias{nodes_subjects}
\title{Create Node Data for Pedigree Subjects}
\usage{
nodes_subjects(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame with studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of nodes for pedigree subjects.
}
\description{
This function builds a node data frame for the pedigree subjects based on metadata,
which can be used for interactive plotting.
}
