% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{nodes_subnucs}
\alias{nodes_subnucs}
\title{Create Node Data for Pedigree Subnuclei}
\usage{
nodes_subnucs(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame of nodes for subnuclei.
}
\description{
This function constructs a node data frame for subnuclei, which serve as connectors
for nuclear family units in the pedigree visualization.
}
