% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{parent_details}
\alias{parent_details}
\title{Retrieve Detailed Parent Information}
\usage{
parent_details(x, studbook)
}
\arguments{
\item{x}{A data frame containing candidate parent records.}

\item{studbook}{A data frame containing the full studbook data.}
}
\value{
A tibble with detailed parent information.
}
\description{
Joins candidate parent records with a condensed version of the studbook to provide detailed information for potential parental assignments. This function is intended for internal use by \code{find_parent}.
}
