% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ped_attribute}
\alias{ped_attribute}
\title{Create a character vector from a studbook dataframe that matches some attribute vector to an individual's sex and status}
\usage{
ped_attribute(df, name, u, m_i, f_i, m_e, f_e, m_d, f_d)
}
\arguments{
\item{df}{A data frame with rowwise location data containing the following columns:
\code{Sex}, \code{exclude}}

\item{name}{A string to use as the name of the new column}

\item{u}{A character vector to assign to the column where \code{Sex} is undetermined}

\item{m_i}{A character vector to assign to the column where \code{Sex} is male and individual is included in the breeding population and therefore alive}

\item{f_i}{A character vector to assign to the column where \code{Sex} is female and individual is included in the breeding population and therefore alive}

\item{m_e}{A character vector to assign to the column where \code{Sex} is male and individual is alive but excluded from the breeding population}

\item{f_e}{A character vector to assign to the column where \code{Sex} is female and individual is alive but excluded from the breeding population}

\item{m_d}{A character vector to assign to the column where \code{Sex} is male and individual is deceased}

\item{f_d}{A character vector to assign to the column where \code{Sex} is female and individual is deceased}
}
\value{
The same data frame with a new character vector column
}
\description{
Create a character vector from a studbook dataframe that matches some attribute vector to an individual's sex and status
}
