% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{ped_coords}
\alias{ped_coords}
\title{Compute Coordinates for Pedigree Nodes Using a Sugiyama Layout}
\usage{
ped_coords(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame containing vertex IDs and corresponding \code{x} and \code{y} coordinates.
}
\description{
This function calculates (x, y) coordinates for nodes in the pedigree igraph
using a Sugiyama layout.
}
