% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_edf}
\alias{ped_edf}
\title{Generate an edge dataframe from the pedigree that functions within \code{DiagrammeR}}
\usage{
ped_edf(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A formatted edge dataframe for \code{DiagrammeR}
}
\description{
Generate an edge dataframe from the pedigree that functions within \code{DiagrammeR}
}
