% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_graphv}
\alias{ped_graphv}
\title{Create a pedigree graph using DiagrammeR}
\usage{
ped_graphv(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
A graphviz object rendered with \code{DiagrammeR}
}
\description{
Create a pedigree graph using DiagrammeR
}
