% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ped_group}
\alias{ped_group}
\title{Add group labels based on a series of metadata attributes to a node dataframe}
\usage{
ped_group(df)
}
\arguments{
\item{df}{A working dataframe with one row per node of a pedigree graph}
}
\value{
The same data frame with two new \code{group} columns (\code{subject_group} and \code{edge_group})
}
\description{
Add group labels based on a series of metadata attributes to a node dataframe
}
