% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{ped_groupIcons}
\alias{ped_groupIcons}
\title{Define Icon Settings for Pedigree Groups}
\usage{
ped_groupIcons()
}
\value{
A list of icon settings.
}
\description{
This function returns a list of FontAwesome icon settings for various pedigree groups.
}
