% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_ids}
\alias{ped_ids}
\title{Generate Internal IDs for Pedigree Members}
\usage{
ped_ids(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame with columns \code{id_subject} and \code{label_subject}.
}
\description{
This function creates a data frame containing internal numeric IDs and corresponding labels for the members of a pedigree.
}
