% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_igraph}
\alias{ped_igraph}
\title{Create a pedigree graph using \code{igraph}}
\usage{
ped_igraph(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
A graph object built with \code{igraph}
}
\description{
Create a pedigree graph using \code{igraph}
}
