% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_levels}
\alias{ped_levels}
\title{Compute Generation and Level Information for a Pedigree}
\usage{
ped_levels(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame with subject IDs, labels, levels, and generation numbers.
}
\description{
This function computes and returns the generation numbers and corresponding levels for each individual
in a pedigree.
}
