% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_ndf}
\alias{ped_ndf}
\title{Create a node data frame to plot a pedigree using DiagrammeR}
\usage{
ped_ndf(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
A formatted node dataframe for graphing with \code{DiagrammeR}
}
\description{
Create a node data frame to plot a pedigree using DiagrammeR
}
