% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_nodes}
\alias{ped_nodes}
\title{Create a complete node data frame that includes connector nodes and original nodes representing individuals}
\usage{
ped_nodes(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of nodes for different visualization tools.
}
\description{
This function creates a template node data frame linking all nodes in the network visualization.
(variables can be selected or removed later depending on which visualization tool you use)
}
