% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{ped_plot_igraph}
\alias{ped_plot_igraph}
\title{Plot a Pedigree igraph Object}
\usage{
ped_plot_igraph(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A base R plot of the pedigree igraph.
}
\description{
This function plots an igraph pedigree using a Sugiyama layout.
}
