% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{ped_verts}
\alias{ped_verts}
\title{Create Vertices Data for an igraph Pedigree Graph}
\usage{
ped_verts(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of vertices with plotting attributes.
}
\description{
This function builds a vertices data frame for constructing an igraph object from pedigree data,
applying custom colors, shapes, and sizes.
}
