% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_visEdges}
\alias{ped_visEdges}
\title{Create a pedigree edge dataframe that works with \code{visNetwork}}
\usage{
ped_visEdges(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
An edge dataframe that works with \code{visNetwork}
}
\description{
Create a pedigree edge dataframe that works with \code{visNetwork}
}
