% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_visGroups}
\alias{ped_visGroups}
\title{Add formatted groups to a \code{visNetwork} pedigree object}
\usage{
ped_visGroups(visNetwork)
}
\arguments{
\item{visNetwork}{A \code{visNetwork} object}
}
\value{
A \code{visNetwork} object with group-based attributes
}
\description{
Add formatted groups to a \code{visNetwork} pedigree object
}
