% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{ped_visIcons}
\alias{ped_visIcons}
\title{Apply Custom Icon Settings to a visNetwork Pedigree Graph}
\usage{
ped_visIcons(graph)
}
\arguments{
\item{graph}{A visNetwork object.}
}
\value{
A modified visNetwork object with custom icon groups applied.
}
\description{
This function applies custom FontAwesome icon settings for various pedigree groups to a visNetwork object.
}
