% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pedigree_births}
\alias{pedigree_births}
\title{Extract and organize tibble of birth events from pedigree object}
\usage{
pedigree_births(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of unique mating pairs with location info and ids to create nodes for a network
}
\description{
Extract and organize tibble of birth events from pedigree object
}
