% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{pedigree_edges}
\alias{pedigree_edges}
\title{Generate edge list for pedigree visNetwork plot}
\usage{
pedigree_edges(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of edges with visual attributes
}
\description{
Generate edge list for pedigree visNetwork plot
}
