% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pedigree_levels}
\alias{pedigree_levels}
\title{Assign generation levels to individuals in a pedigree}
\usage{
pedigree_levels(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object from \code{pedtools::ped()}}
}
\value{
A tibble of individual IDs and generation levels
}
\description{
Assign generation levels to individuals in a pedigree
}
