% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{pedNode_visGroup}
\alias{pedNode_visGroup}
\title{Set Custom Group Appearance for a Specific Node Group in a visNetwork Graph}
\usage{
pedNode_visGroup(graph, groupname)
}
\arguments{
\item{graph}{A visNetwork object.}

\item{groupname}{A character string representing the group name.}
}
\value{
The visNetwork object with the specified group settings applied.
}
\description{
This function assigns a custom image to a specific node group in a visNetwork graph.
}
