% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{pedsave}
\alias{pedsave}
\title{Create download a local copy of a pedigree plot as a png file}
\usage{
pedsave(pedigree, studbook, name = NULL, palette = NULL, labs_spec = FALSE)
}
\arguments{
\item{pedigree}{A single pedigree object (from \code{pedtools::ped()})}

\item{studbook}{A tibble with \code{ID}, \code{Sex}, \code{Status}, and other metadata}

\item{name}{A title for the plot (used as label)}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
An html file containing a rendered pedigree plot saved to a local subdirectory called \code{zoopop_plots}
}
\description{
Create download a local copy of a pedigree plot as a png file
}
