% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{pedsave_series}
\alias{pedsave_series}
\title{Create download a local copy of a pedigree plot as a png file}
\usage{
pedsave_series(pedigree_series, studbook, palette = NULL, labs_spec = FALSE)
}
\arguments{
\item{pedigree_series}{A list of pedigree objects.}

\item{studbook}{A tibble with \code{ID}, \code{Sex}, \code{Status}, and other metadata}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
An html file containing a rendered pedigree plot saved to a local subdirectory called \code{zoopop_plots}
}
\description{
Create download a local copy of a pedigree plot as a png file
}
