% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_age_ts}
\alias{plot_age_ts}
\title{Plot age-specific demographic variables with animation by birth year and sex by line color}
\usage{
plot_age_ts(studbook, cohort_params = NULL, variable, log_trans = FALSE)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{variable}{Name of the column containing an age-specific variable to map onto the y-axis
options for variable: \code{Births}, \code{Deaths}, \code{Nx}, \code{Qx}, \code{Lx}, \code{Lx1}, \code{Px}, \code{ex}, \code{Tx}, \code{Mx}, \code{Fx}}

\item{log_trans}{Logical indicating whether to log-transform the y-axis variable}
}
\value{
A plotly object
}
\description{
Plot age-specific demographic variables with animation by birth year and sex by line color
}
