% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_census}
\alias{plot_census}
\title{Plot population trends based on census data calculated from studbook}
\usage{
plot_census(studbook)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}
}
\value{
A plotly plot object
}
\description{
Plot population trends based on census data calculated from studbook
}
