% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog}
\alias{plot_demog}
\title{Plot lifetime demographic variables by sex and birth year}
\usage{
plot_demog(studbook, cohort_params = NULL, variable)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{variable}{Name of the column containing an age-specific variable to map onto the y-axis
options for variable: \code{N0}, \code{N1}, \code{R0}, \code{T}, \code{MLE}, \code{Repro_first}, \code{Repro_last}, \code{age_max}, \code{lambda}, \code{r}}
}
\value{
A plotly object
}
\description{
Plot lifetime demographic variables by sex and birth year
}
