% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog_annot}
\alias{plot_demog_annot}
\title{Generate a list of shapes and annotations to use in plotly plots of demographic variables}
\usage{
plot_demog_annot(attr, overall_val = NULL, age_spec = TRUE, log_trans = FALSE)
}
\arguments{
\item{attr}{A list of attribute values created using the internal function \code{plot_demog_attr}}

\item{overall_val}{value of y when all cohorts are combined (\code{NULL} when plotting age-specific variables)}

\item{age_spec}{Logical indicating whether to create list for age-specific version or not}

\item{log_trans}{Logical indicating whether to log-transform the y-axis variable}
}
\value{
A list of parameter values to use in plotly plots of demographic variables
}
\description{
Generate a list of shapes and annotations to use in plotly plots of demographic variables
}
\keyword{internal}
