% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog_attr}
\alias{plot_demog_attr}
\title{Generate a list of attributes to use in plotly plots of demographic variables}
\usage{
plot_demog_attr(data, variable, age_spec = TRUE)
}
\arguments{
\item{data}{A dataframe to use in plotly (generated internally using the function \code{plot_demog_prep})}

\item{variable}{Name of the column containing a variable to map onto the y-axis
options for age-specific variable: \code{Births}, \code{Deaths}, \code{Nx}, \code{Qx}, \code{Lx}, \code{Lx1}, \code{Px}, \code{ex}, \code{Tx}, \code{Mx}, \code{Fx}
options for lifetime variable: \code{N0}, \code{N1}, \code{R0}, \code{T}, \code{MLE}, \code{Repro_first}, \code{Repro_last}, \code{age_max}, \code{lambda}, \code{r}}

\item{age_spec}{Logical indicating whether to create list for age-specific version or not}
}
\value{
A list of parameter values to use in plotly plots of demographic variables
}
\description{
Generate a list of attributes to use in plotly plots of demographic variables
}
