% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog_layout}
\alias{plot_demog_layout}
\title{Add a layout to a demographic plotly object}
\usage{
plot_demog_layout(plot, layer, attr, age_spec = TRUE)
}
\arguments{
\item{plot}{A plotly object}

\item{layer}{A list of shapes and text annotations created using the internal function \code{plot_demog_annot}}

\item{attr}{A list of attribute values created using the internal function \code{plot_demog_attr}}

\item{age_spec}{Logical indicating whether the y-variable plotted is age-specific or not}
}
\value{
A list of parameter values to use in plotly plots of demographic variables
}
\description{
Add a layout to a demographic plotly object
}
