% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog_prep}
\alias{plot_demog_prep}
\title{Prepare demographic tibble for plotting with plotly}
\usage{
plot_demog_prep(
  studbook,
  cohort_params = NULL,
  variable,
  log_trans = FALSE,
  age_spec = TRUE,
  sex = NULL
)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{variable}{Name of the column containing an age-specific variable to map onto the y-axis
age-specific options for variable: \code{Births}, \code{Deaths}, \code{Nx}, \code{Qx}, \code{Lx}, \code{Lx1}, \code{Px}, \code{ex}, \code{Tx}, \code{Mx}, \code{Fx}
options for lifetime variable: \code{N0}, \code{N1}, \code{R0}, \code{T}, \code{MLE}, \code{Repro_first}, \code{Repro_last}, \code{age_max}, \code{lambda}, \code{r}}

\item{log_trans}{Logical indicating whether to log-transform the y-axis variable}

\item{age_spec}{Logical indicating whether to create list for age-specific version or not}

\item{sex}{One of \code{"Males"}, \code{"Females"}, or \code{"Overall"} to plot stats by sex (\code{NULL} produces a dataframe with all three categories)}
}
\value{
A dataframe to use in plotly
}
\description{
Prepare demographic tibble for plotting with plotly
}
