% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_demog_trace}
\alias{plot_demog_trace}
\title{Add a scatter trace for a demographic variable to a plotly object}
\usage{
plot_demog_trace(
  plot,
  data,
  attr,
  primary = TRUE,
  age_spec = TRUE,
  sex = c("Males", "Females", "Overall")
)
}
\arguments{
\item{plot}{A plotly object}

\item{data}{A dataframe to use in plotly (generated internally using the function \code{plot_demog_prep})}

\item{attr}{A list of attribute values created using the internal function \code{plot_demog_attr}}

\item{primary}{Logical indicating whether the trace will be primary in the plot (\code{TRUE}) or a background trace (\code{FALSE})}

\item{age_spec}{Logical indicating whether the y-variable plotted is age-specific or not}

\item{sex}{Vector of sex classes to include in trace (default is all three)}
}
\value{
A list of parameter values to use in plotly plots of demographic variables
}
\description{
Add a scatter trace for a demographic variable to a plotly object
}
