% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambda growth rates by cohort}
\usage{
plot_lambda(studbook, cohort_params = NULL, caption = NULL, number = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{caption}{The caption text for the plot (optional)}

\item{number}{The optional number to add to the plot caption (e.g., Figure 1.)}
}
\value{
A plotly object
}
\description{
Plot lambda growth rates by cohort
}
