% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{proj_input}
\alias{proj_input}
\title{Add projected values to demographic summary table}
\usage{
proj_input(studbook, cohort_params = NULL, sex = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{sex}{Which group to include for summary under the \code{Sex} column - one of \code{Males}, \code{Females}, or \code{Overall} (optional, default is \code{Overall})}
}
\value{
Summary table with projection values to be used in leslie matrix construction
}
\description{
Add projected values to demographic summary table
}
