% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{proj_matrix_overall}
\alias{proj_matrix_overall}
\title{Add projection matrices to demographic summary table with all birth-year cohorts merged}
\usage{
proj_matrix_overall(studbook, iterations = 20)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{iterations}{Number of year-steps to project out through iterations (default = 20 years)}
}
\value{
Summary table with projection values factoring in all individuals across all birth years
}
\description{
Add projection matrices to demographic summary table with all birth-year cohorts merged
}
