% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{proj_ts}
\alias{proj_ts}
\title{Wrapper function to compute grid and ts columns for all age-specific variables}
\usage{
proj_ts(proj_df, variables = c("Fx", "Mx", "Qx", "Px", "Tx", "ex"))
}
\arguments{
\item{proj_df}{Summary table with projection values (produced by \code{proj_matrix})}

\item{variables}{Names of the variables for which the values should be computed (default is c(\code{Fx},\code{Mx},\code{Qx},\code{Px},\code{Tx},\code{ex}))}
}
\value{
Summary table with projection time series list-columns for values across all ages projected for a given variable
}
\description{
Wrapper function to compute grid and ts columns for all age-specific variables
}
