% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{proj_var_grid}
\alias{proj_var_grid}
\title{Compute a full time×age grid for a given variable at each t (*_grid) for a projection summary table}
\usage{
proj_var_grid(proj_df, var)
}
\arguments{
\item{proj_df}{Summary table with projection values (produced by \code{proj_matrix})}

\item{variable}{Name of the variable for which the time series should be projected
\itemize{
\item one of: \code{Fx} (reproductive output), \code{Mx} (fertility rate), \code{Qx} (mortality rate), \code{Px} (survival probability), \code{Tx} (total future lifetime), \code{ex} (life expectancy)
}}
}
\value{
Summary table with projection time series list-columns for values across all ages projected for a given variable
}
\description{
Compute a full time×age grid for a given variable at each t (*_grid) for a projection summary table
}
