% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections_overall_studbook}
\alias{projections_overall_studbook}
\title{Create a detailed projection summary table from a studbook tibble with all birth-year cohorts merged}
\usage{
projections_overall_studbook(studbook, iterations = 20)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{iterations}{Number of year-steps to project out through iterations (default = 20 years)}
}
\value{
Summary table with projection values, including full \verb{time x age} grids and total time-series list-cols for each age-specific variable
}
\description{
Create a detailed projection summary table from a studbook tibble with all birth-year cohorts merged
}
