% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections_studbook}
\alias{projections_studbook}
\title{Create a detailed projection summary table from a studbook tibble}
\usage{
projections_studbook(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
Summary table with projection values, including full \verb{time x age} grids and total time-series list-cols for each age-specific variable
}
\description{
Create a detailed projection summary table from a studbook tibble
}
