% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{ID}
\alias{ID}
\alias{exclude}
\alias{Date_birth}
\alias{Date_last}
\alias{Loc_birth}
\alias{Loc_last}
\alias{age_last}
\alias{bubble_count}
\alias{Sire}
\alias{Dam}
\alias{Rel_Contribution}
\alias{inbred}
\title{Reactable Column Defs}
\usage{
ID(df)

exclude(df)

Date_birth(df, colors = colors)

Date_last(df, colors = colors)

Loc_birth(df)

Loc_last(df)

age_last(df, colors = colors)

bubble_count(df, name, colors = colors)

Sire(df, colors = colors)

Dam(df, colors = colors)

Rel_Contribution(df, colors = colors)

inbred(df, colors = colors)
}
\arguments{
\item{df}{A data frame passed to the reactable column definition}

\item{name}{The name of the column to be shown in the table}
}
\description{
These helper functions generate formatted column definitions for \code{reactable}
tables to summarize studbook, pedigree, and demographic data.
}
