% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{read_btp}
\alias{read_btp}
\title{Read and Process BTP Data}
\usage{
read_btp(file_in, file_out, loc_key)
}
\arguments{
\item{file_in}{A file path to the input CSV file created by converting a published BTP to a CSV table. The file must include the columns: \code{BTP}, \code{Institution}, \code{SB ID}, \code{Sex}, \code{Age}, \code{Disposition}, \code{Location}, \code{Breeding}, \code{With}, \code{Notes}, and \code{Facility_Note}. (You should add the \code{BTP} column manually to indicate the 4-digit publication year.)}

\item{file_out}{A file path to the output CSV file that will contain the processed BTP summary.}

\item{loc_key}{A data frame containing location key mappings produced by \code{read_locations}.}
}
\value{
A data frame with processed BTP data.
}
\description{
Reads a CSV file containing Breeding and Transfer Plan (BTP) data, processes the data, writes an updated BTP summary to an output CSV file, and returns the resulting data frame.
}
