% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_studbook.R
\name{read_studbook}
\alias{read_studbook}
\title{Read and Process Studbook Data}
\usage{
read_studbook(file_in, loc_key, btp, names = NULL, add_births = NULL)
}
\arguments{
\item{file_in}{A file path to the input CSV file. Export the studbook tables from the published PDF using Adobe Acrobat (or a similar tool), and ensure that the file contains the columns: \code{ID}, \code{Sex}, \code{Sire}, \code{Dam}, \code{Birth_Type}, \code{Event}, \code{Date}, and \code{Location}. Combine rows from multiple files as needed.}

\item{loc_key}{A data frame containing location key mappings produced by \code{read_locations}.}

\item{btp}{A data frame containing formatted BTP data produced by \code{read_btp}.}

\item{names}{Factor labeling vector written as \verb{c(\{String\} = "\{ID\}")} for assigning special name labels to selected ID numbers (optional - will return empty cells for all rows if not included).}

\item{add_births}{List provided to account for new births implied but not documented in most recent studbook. The top level list names should be an estimated date of birth. The next level should be the names \code{ID}, \code{Sire}, and \code{Dam} with a numeric ID value provided for each. (if \code{add_births} is not used, then this option is ignored)}
}
\value{
A data frame with processed studbook data that can be used to resolve missing parental assignments.
}
\description{
Reads a CSV file containing studbook data, processes it together with BTP data and location key mappings, and returns a cleaned studbook data frame. This function merges multiple exported files, removes redundant entries, and performs necessary corrections for downstream analyses.
}
