% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetab.R
\name{repro_history}
\alias{repro_history}
\title{Generate a summary of reproductive events across population history}
\usage{
repro_history(studbook, window = NULL)
}
\arguments{
\item{window}{vector of start and end date (must be provided as \code{ymd()} values) to use as the demographic window for calculations
(default is the date of the first captive birth recorded and the most recent date of birth recorded)}

\item{stubook}{tibble organized and formatted originally using \code{read_studbook}}
}
\value{
Tibble summarizing the reproductive events across studbook records within the demographic window
(\code{ID} of deceased individual, \code{Sex} of individual, \code{Type_birth} (represents the parent/ID, not offspring) indicating \code{Wild} for wild captures or
\code{Captive} for captive-born individuals, \code{Qtr_born} is the date used to represent birth date of the parent/ID - which are converted to the starting date of the annual quarter,
\code{x_birth} is the approximate age class in years in which the individual reproduces based on the
difference between the start of the quarter of their birth and the start of the quarter of their offspring's birth,
\code{sex_birth} is the sex of the offspring)
}
\description{
Generate a summary of reproductive events across population history
}
