% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_demog.R
\name{rows_bySex}
\alias{rows_bySex}
\title{Style each row of a reactable demographic table based on groups by Sex}
\usage{
rows_bySex()
}
\description{
Style each row of a reactable demographic table based on groups by Sex
}
