% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_markers}
\alias{set_markers}
\title{Generate symbol mapping for plotly plots}
\usage{
set_markers(f = "circle", m = "square", u = "diamond", t = "cross")
}
\arguments{
\item{f}{A single symbol to map to females}

\item{m}{A single symbol to map to males}

\item{u}{A single symbol to map to undetermined sex}

\item{t}{A single symbol to map to totals}
}
\value{
A named list of symbols to reference in plotly functions and plots
}
\description{
If you do not wish to assign custom marker symbols, this will apply default symbols to your named list that are referenced in other functions in this package.
}
