% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{set_ped_paths}
\alias{set_ped_paths}
\title{Create a data frame with labels and unique numeric ids to use for building nodes and edges of a pedigree diagram}
\usage{
set_ped_paths(studbook, pedigree, palette = NULL)
}
\arguments{
\item{pedigree}{A pedigree object}
}
\value{
A new dataframe to use for building edge and node dataframes with \code{DiagrammeR}
}
\description{
Create a data frame with labels and unique numeric ids to use for building nodes and edges of a pedigree diagram
}
