% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_plotly_pal}
\alias{set_plotly_pal}
\title{Convert a named list palette into named vector format}
\usage{
set_plotly_pal(palette)
}
\arguments{
\item{palette}{A named list of color values}
}
\value{
A named vector suitable for plotly or reactable inputs
}
\description{
Convert a named list palette into named vector format
}
