% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{special_cohorts}
\alias{special_cohorts}
\title{Create Special Cohort Labels based on values in \code{name_spec}}
\usage{
special_cohorts(
  studbook,
  Year_min = NULL,
  Year_max = NULL,
  span = 10,
  age_max = NULL
)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{Year_min}{Start year of cohorts (optional)}

\item{Year_max}{End year (optional)}

\item{span}{Years per cohort (optional)}

\item{age_max}{Max age to include (optional)}
}
\value{
A joined and restructured tibble
}
\description{
Create Special Cohort Labels based on values in \code{name_spec}
}
