% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{studbook.cols}
\alias{studbook.cols}
\title{Column definitions for full studbook summary tables}
\usage{
studbook.cols(df)
}
\arguments{
\item{df}{A filtered founder summary dataframe}
}
\value{
A named list of column definitions
}
\description{
Column definitions for full studbook summary tables
}
