% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{studbook_cols}
\alias{studbook_cols}
\title{Column definitions for full studbook summary tables}
\usage{
studbook_cols(df, df_cols, colors = colors)
}
\arguments{
\item{df}{A filtered founder summary dataframe}
}
\value{
A named list of column definitions
}
\description{
Column definitions for full studbook summary tables
}
