% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{summarize_subnucs}
\alias{summarize_subnucs}
\title{Summarize Subnuclear Units in a Pedigree}
\usage{
summarize_subnucs(pedigree)
}
\arguments{
\item{pedigree}{A pedigree object.}
}
\value{
A data frame summarizing each subnucleus with the number of children.
}
\description{
This function groups the subnuclei (nuclear family units) and counts the number
of offspring within each.
}
