% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{visPed}
\alias{visPed}
\title{Build an Interactive visNetwork Pedigree Graph}
\usage{
visPed(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A visNetwork object representing the pedigree graph.
}
\description{
This function creates an interactive pedigree graph using visNetwork by combining nodes,
edges, and custom icon settings.
}
