% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{visPed_interaction}
\alias{visPed_interaction}
\title{Apply Custom Interaction Options to a visNetwork Pedigree Graph}
\usage{
visPed_interaction(graph)
}
\arguments{
\item{graph}{A visNetwork object.}
}
\value{
A modified visNetwork object with custom options.
}
\description{
This function applies multiple predefined interaction settings for various pedigree groups to a visNetwork object.
}
