% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{visPed_options}
\alias{visPed_options}
\title{Apply Custom Visualization Options to a visNetwork Pedigree Graph}
\usage{
visPed_options(graph, variable = NULL)
}
\arguments{
\item{graph}{A visNetwork object.}
}
\value{
A modified visNetwork object with custom options.
}
\description{
This function applies multiple predefined options settings for various pedigree groups to a visNetwork object.
}
