% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{visPed_tree}
\alias{visPed_tree}
\title{Create a Hierarchical visNetwork Pedigree Tree}
\usage{
visPed_tree(studbook, pedigree, variable = NULL)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A visNetwork object with a hierarchical layout applied.
}
\description{
This function applies a hierarchical layout to a visNetwork pedigree graph.
}
