% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{build_ped_series}
\alias{build_ped_series}
\title{Build a Series of Pedigree Objects from a Studbook}
\usage{
build_ped_series(studbook)
}
\arguments{
\item{studbook}{A data frame containing studbook data.}
}
\value{
A list of pedigree objects.
}
\description{
This function converts a studbook into a list of pedigree objects using
\code{pedtools::ped()} and discards any singletons (pedigrees with size one or less).
}
