% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{cohort_lifeTab}
\alias{cohort_lifeTab}
\title{Generate life table from studbook over cohorts}
\usage{
cohort_lifeTab(
  studbook,
  Year_min = NULL,
  Year_max = NULL,
  span = 10,
  age_max = NULL
)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{Year_min}{First cohort year (optional)}

\item{Year_max}{Last cohort year (optional)}

\item{span}{Years per cohort (optional)}

\item{age_max}{Maximum age (optional)}
}
\value{
A tibble with cohort life table summary
}
\description{
Generate life table from studbook over cohorts
}
