% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_births}
\alias{count_births}
\title{Count births and track individual presence over time}
\usage{
count_births(studbook)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}
}
\value{
A tibble of births per ID by year
}
\description{
Count births and track individual presence over time
}
