% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{extract_pedigree}
\alias{extract_pedigree}
\title{Extract a Specific Pedigree from a Pedigree Series}
\usage{
extract_pedigree(pedigree_series, name = NULL, title = NULL)
}
\arguments{
\item{pedigree_series}{A list of pedigree objects.}

\item{name}{Optional. The name of the pedigree object to extract.}

\item{title}{Optional. A new title to assign to the extracted pedigree.}
}
\value{
A single pedigree object.
}
\description{
This function extracts a pedigree object from a list of pedigrees based on a provided name.
If no name is supplied, the pedigree with maximum depth is chosen.
}
