% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loc_tooltip}
\alias{loc_tooltip}
\title{Create a character vector from a studbook dataframe to represent location data in a tooltip}
\usage{
loc_tooltip(df, name)
}
\arguments{
\item{df}{A data frame with rowwise location data containing the columns \code{iconLoc}, \code{Loc}, \code{Institution}, and \code{State_Province}}

\item{name}{A string to use as the name of the new column}
}
\value{
The same data frame with a new string column
}
\description{
Create a character vector from a studbook dataframe to represent location data in a tooltip
}
