% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_cohorts}
\alias{make_cohorts}
\title{Generate cohort-structured data}
\usage{
make_cohorts(
  studbook,
  Year_min = NULL,
  Year_max = NULL,
  span = 10,
  age_max = NULL,
  include_sex = TRUE
)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{Year_min}{Start year of cohorts (optional)}

\item{Year_max}{End year (optional)}

\item{span}{Years per cohort (optional)}

\item{age_max}{Max age to include (optional)}

\item{include_sex}{Whether to include sex as a grouping var (optional)}
}
\value{
A joined and restructured tibble
}
\description{
Generate cohort-structured data
}
