% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{make_cohorts}
\alias{make_cohorts}
\title{Generate cohort-structured data}
\usage{
make_cohorts(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}

\item{include_sex}{Whether to include sex as a grouping var (optional)}
}
\value{
A joined and restructured tibble
}
\description{
Generate cohort-structured data
}
