% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_edges}
\alias{ped_edges}
\title{Create a data frame with basic metadata attached to edges connecting individuals}
\usage{
ped_edges(studbook, pedigree, palette = NULL)
}
\arguments{
\item{studbook}{A studbook tibble}

\item{pedigree}{A pedigree object}

\item{palette}{An optional named list mapping individuals to colors by sex and
status; if not provided, \code{\link[=set_colors]{set_colors()}} is used.}
}
\value{
A new dataframe to use for building edge and node dataframes with \code{DiagrammeR}
}
\description{
Create a data frame with basic metadata attached to edges connecting individuals
}
