% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_edges}
\alias{ped_edges}
\title{Combine Pedigree Edge Data}
\usage{
ped_edges(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of combined edges.
}
\description{
This function combines the edges for children, subnuclei, maternal, and paternal connections
into a single edge data frame for pedigree visualization.
}
