% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{ped_igraph}
\alias{ped_igraph}
\title{Build an igraph Object from Pedigree Data}
\usage{
ped_igraph(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
An igraph object representing the pedigree.
}
\description{
This function creates an igraph object from pedigree edge and vertex data.
}
