% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{ped_ndf}
\alias{ped_ndf}
\title{Generate a node dataframe from the pedigree that functions within \code{DiagrammeR}}
\usage{
ped_ndf(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A formatted node dataframe for \code{DiagrammeR}
}
\description{
Generate a node dataframe from the pedigree that functions within \code{DiagrammeR}
}
