% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{ped_visEdges}
\alias{ped_visEdges}
\title{Create Edge Data for visNetwork Pedigree Graphs}
\usage{
ped_visEdges(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of edges formatted for visNetwork.
}
\description{
This function constructs an edge data frame for a pedigree graph formatted for visNetwork.
}
