% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{ped_visGroups}
\alias{ped_visGroups}
\title{Apply Custom Group Settings to a visNetwork Pedigree Graph}
\usage{
ped_visGroups(graph)
}
\arguments{
\item{graph}{A visNetwork object.}
}
\value{
A modified visNetwork object with custom group appearances.
}
\description{
This function applies multiple predefined group settings for various pedigree groups to a visNetwork object.
}
