% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{ped_visNodes}
\alias{ped_visNodes}
\title{Create Node Data for visNetwork Pedigree Graphs}
\usage{
ped_visNodes(studbook, pedigree)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{pedigree}{A pedigree object.}
}
\value{
A data frame of nodes formatted for visNetwork.
}
\description{
This function builds a node data frame for rendering a pedigree graph using visNetwork,
combining subject and subnucleus nodes.
}
