% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_base_plot.R
\name{pedsave}
\alias{pedsave}
\title{Save a Pedigree Plot to HTML}
\usage{
pedsave(pedigree, studbook, palette = NULL, labs_spec = FALSE)
}
\arguments{
\item{pedigree}{A pedigree object.}

\item{studbook}{A data frame containing studbook metadata.}

\item{palette}{Optional. A named list mapping individuals to colors.}

\item{labs_spec}{Logical. If \code{TRUE}, emphasizes individuals with special labels.}

\item{name}{Optional. The title for the plot; defaults to the pedigree family ID.}
}
\value{
None. The function saves an HTML file.
}
\description{
This function saves a pedigree plot as an HTML file in the local directory \code{zoopop_plots}.
}
