% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_census}
\alias{plot_census}
\title{Plot population trends using census data}
\usage{
plot_census(studbook, colors = NULL, title = NULL)
}
\arguments{
\item{studbook}{A data frame containing studbook metadata.}

\item{colors}{A named list of colors for "f", "m", and "u" (optional)}

\item{title}{The title for the plot (optional)}
}
\value{
A plotly plot object
}
\description{
Plot population trends using census data
}
