% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambda growth rates by cohort}
\usage{
plot_lambda(life_table, colors = NULL, title = NULL)
}
\arguments{
\item{life_table}{A tibble with lambda values and Sex, Cohort_label columns}

\item{title}{Plot title}

\item{palette}{A color palette in list format}
}
\value{
A plotly object
}
\description{
Plot lambda growth rates by cohort
}
