% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_demog.R
\name{rct_demog}
\alias{rct_demog}
\title{Create a reactable summary table of demographic stats using a formatted studbook}
\usage{
rct_demog(studbook, cohort_params = NULL)
}
\description{
Create a reactable summary table of demographic stats using a formatted studbook
}
