% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{ID}
\alias{ID}
\alias{Status}
\alias{exclude}
\alias{Loc_birth}
\alias{Loc_last}
\alias{bubble_count}
\alias{Sire}
\alias{Dam}
\alias{Rel_Contribution}
\alias{inbred}
\title{Reactable Column Defs}
\usage{
ID(df)

Status()

exclude(df, colors = set_colors())

Loc_birth(df)

Loc_last(df)

bubble_count(df, name, colors = set_colors())

Sire(df, colors = set_colors())

Dam(df, colors = set_colors())

Rel_Contribution(df, colors = set_colors())

inbred(df, colors = set_colors())
}
\arguments{
\item{df}{A data frame passed to the reactable column definition}

\item{name}{The name of the column to be shown in the table}
}
\description{
These helper functions generate formatted column definitions for \code{reactable}
tables to summarize studbook, pedigree, and demographic data.
}
