% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldefs.R, R/reactables.R
\name{Cohort_years}
\alias{Cohort_years}
\alias{Sex}
\alias{icon_cells}
\alias{sparkline_cells}
\alias{intrins_rate}
\alias{n1}
\alias{nx}
\alias{lx1}
\alias{px}
\alias{qx}
\alias{MLE}
\alias{tx}
\alias{ex}
\alias{age_ranges}
\alias{gen_T}
\alias{mx}
\alias{fx}
\alias{r0}
\alias{ID}
\alias{Status}
\alias{exclude}
\alias{Loc_birth}
\alias{Loc_last}
\alias{bubble_count}
\alias{Sire}
\alias{Dam}
\alias{Rel_Contribution}
\alias{inbred}
\title{Reactable Column Defs}
\usage{
Cohort_years()

Sex()

icon_cells(data, seq_by)

sparkline_cells(data)

intrins_rate()

n1(data)

nx(data)

lx1(data)

px(data)

qx(data)

MLE(data)

tx(data)

ex(data)

age_ranges(data)

gen_T(data)

mx(data)

fx(data)

r0(data)

ID(df)

Status()

exclude(df, colors)

Loc_birth(df)

Loc_last(df)

bubble_count(df, name, colors)

Sire(df, colors)

Dam(df, colors)

Rel_Contribution(df, colors)

inbred(df, colors)
}
\arguments{
\item{data}{Dataframe passed to reactable table function}

\item{seq_by}{A numerical input that determines what number each icon represents.}

\item{df}{A data frame passed to the reactable column definition}

\item{colors}{A named list of colors to use (must a list of colors named \code{seq}, a list of colors named \code{div}, one color for \code{f} and one for \code{m})}

\item{name}{The name of the column to be shown in the table}
}
\description{
These helper functions generate formatted column definitions for \code{reactable}
tables to summarize studbook, pedigree, and demographic data.
}
