% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{ID}
\alias{ID}
\alias{Status}
\alias{DateBirth}
\alias{DateDeath}
\alias{LocBirth}
\alias{LocLast}
\alias{Current_Location}
\alias{AgeLast}
\alias{bubble_count}
\alias{Sire}
\alias{Dam}
\alias{Rel_Contribution}
\alias{inbred}
\title{Reactable Column Defs}
\usage{
ID(df)

Status()

DateBirth(df)

DateDeath(df)

LocBirth(df)

LocLast(df)

Current_Location(df)

AgeLast(df)

bubble_count(df, name)

Sire(df)

Dam(df)

Rel_Contribution(df)

inbred(df)
}
\arguments{
\item{df}{A data frame passed to the reactable column definition}

\item{name}{The name of the column to be shown in the table}
}
\description{
These helper functions generate formatted column definitions for \code{reactable}
tables to summarize studbook, pedigree, and demographic data.
}
