% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_colors}
\alias{set_colors}
\title{Generate color mapping for reactable tables}
\usage{
set_colors(
  seq = NULL,
  div = NULL,
  rnd = NULL,
  emp = "#B24422FF",
  f = "#C44D76FF",
  m = "#4457A5FF",
  u = "#7CAF5CFF",
  t = "#59386CFF"
)
}
\arguments{
\item{seq}{A named list of color values to use for sequential mapping}

\item{div}{A named list of color values to use for divergent mapping}

\item{rnd}{A named list of color values to use for randomized, qualitative mapping}

\item{emp}{A single color to map to emphasized values}

\item{f}{A single color to map to females}

\item{m}{A single color to map to males}

\item{u}{A single color to map to undetermined sex}

\item{sire}{A single color to map to sires}

\item{dam}{A single color to map to dams}
}
\value{
A named list of color palettes to reference in reactable functions and plots
}
\description{
If you do not wish to assign custom color values, this will apply default colors to your named list that are referenced in other functions in this package.
}
