% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog.R
\name{special_cohorts}
\alias{special_cohorts}
\title{Create Special Cohort Labels based on values in \code{name_spec}}
\usage{
special_cohorts(studbook, cohort_params = NULL)
}
\arguments{
\item{studbook}{A data frame of studbook data produced by \code{read_studbook}.}

\item{cohort_params}{A named list of the parameter values to use for cohorts (\code{Year_min}, \code{Year_max}, \code{span}, \code{age_max}) (optional)}
}
\value{
A joined and restructured tibble
}
\description{
Create Special Cohort Labels based on values in \code{name_spec}
}
