% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{studbook_cols}
\alias{studbook_cols}
\title{Column definitions for full studbook summary tables}
\usage{
studbook_cols(df, df_cols, colors = NULL)
}
\arguments{
\item{df}{A filtered founder summary dataframe}

\item{colors}{A named list of colors to use (optional - must a list of colors named \code{seq}, a list of colors named \code{div}, one color for \code{f} and one for \code{m})}
}
\value{
A named list of column definitions
}
\description{
Column definitions for full studbook summary tables
}
