% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactables.R
\name{studbook_react}
\alias{studbook_react}
\title{Create a reactable summary table from any dataset}
\usage{
studbook_react(df, cols, colors = set_colors(), ...)
}
\arguments{
\item{df}{The data frame to summarize}

\item{cols}{A named list of column definitions}

\item{...}{Additional arguments passed to \code{reactable()}}
}
\value{
A reactable widget
}
\description{
Create a reactable summary table from any dataset
}
